core/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.20
botocore==1.23.20
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-rest-auth==2.2.2
Django==3.2.9
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-extensions==3.1.5
django-sort-order-field==1.2
django-storages==1.11.1
django-tenant-users==0.3.12
django-tenants==3.3.4
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==4.7.2
drf-writable-nested==0.6.3
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.2
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.23
psycopg2==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
